requirements.txt
customtkinter
requests
Pillow
ollama
langchain
sentence-transformers
faiss-cpu
python-dotenv